**Improve SQL query for stats page**
Use a better query to retrieve repositories with the most releases.

⚡️  Goes from 3s to 100ms. 💨 

**Use same query for RSS & dashboard**
Instead of grouping repo to display only the latest release on the dashboard, use the same query for the RSS which will display all latest releases non-grouping per repo.
This speedup the query and the dashboard display.

⚡️ Goes from 2s to 55ms. 💨